requirements.txt
langchain-community
langchain-openai
faiss-cpu
python-dotenv
langchain
langchain-core
openai
streamlit